Pre-dev (#64)
* credential error. need to fix. was working

* use default.config

* dynamodb resource added

* borrow checkergit branchgit branch

* borrow checker battle

* hits upload object but send needs to be async

* data streams to bucket

* ep-dev-a before checkout

* quick save

* merge conflict fixing

* batch process the datastream

* todo - potentially pass in a buffer

* get the buffers passed into the log handler

* added data buffer cleanup on termination signal

* next - calc buffer size and send

* next - calc buffer size and send

* next - calc buffer size and send

* next - calc buffer size and send

* workonbatching

* workonbatching

* do-dynamo

* ep-dynamo-working

* Made the following changes:

Cargo.toml:
	- Added clap crate for experimenting cmnd line args
main.rs:
	- Added cmnd line functionality for specifying destinations
		User needs to specify what destination to send the log
		files now; will work more on this

* remove_dynamodb/add.rs_and_cleanup_startup_logic

* update

* update default config

* update and rename branch

* first

* update src/config.rs

* trait implemented with simpler interface for handle_log_data

* trait implemented with simpler interface for handle_log_data

* sending to dynamo

* changebranchafteraddingopensearch

* ssddforarc

* toml config cleaned up setup

* toml config cleaned up setup

* toml config cleaned up setup

* toml config cleaned up setup

* toml config cleaned up setup

* deleted error.rs. it somehow lived this long

* Jicxer dynamo cmdline (#46)

* Added utility functions

* Moved exit process

* rebased ep-kdf-a and moved dynamodb functionality to util.rs

* Modifications in file iam.rs and main.rs

main.rs:
	Creating the iam client
iam.rs:
	starting iam client and grabbing user policies

* Changes with this commit;

Cargo.toml:
	Added aws-sdk-iam dependency
scr/iam.rs:
	Restarting from scratch to start iam client.
	Aim to grab:
		- Username
		- Attached policies
		- List the possible destinations based on policies
		- Add policies if needed
src/main.rs
	Added a line to start iam client

* update

* update default config

* update and rename branch

* Added utility functions

* Moved exit process

* Working on IAM policy for users
Modified files:

Cargo.toml:
	Added aws-sdk-iam and async-trait
README.md:
	Added resources for AWS SDK IAM
src/iam.rs:
	Created a function to eventually list users
src/main.rs:
	Linked utility function to start IAM

* Worked further on IAM implementation. Added comments to brainstorm
solutions and thinking process.

Modifications

src/iam.rs:
	Implemented the delete user and list user. Functions now work.
src/util.rs:
	Implemented a scenario where a testuser is created, list all the
current users in the IAM, delete the user, check the list.

* Added functionalities for entering AWS CLI based on user groups

Modified:
src/iam.rs:
	- Added list_groups_for_user to check the current group policies
	  that user is in.
	- Added is_admin_user. Checks if the current user contains the
	  group policy "admin". Otherwise, return false
	- get_user to get the current user and store info in
	  GetUserOutput
	-
src/main.rs:
	- Calling run_admin_cli
src/util.rs:
	- Created fmt::Display trait to allow string formatting for
	  getting the user information from get_user
	- Implemented run_admin_cli to run AWS CLI if user is within the
	  "admin" group policy. Otherwise, deny them access and output a
		message
testuser-cred.txt:
	- Credential information for the test-user. Contains "user"
		group policy with ReadOnly Access for IAM, KDF, and
		DynamoDB. Also contains inline user-policy to insert
		contents and create tables for DynamoDB.
		(Will need to look into it as I have not created a table
to test this yet)

* Update testuser-cred.txt

* Delete testuser-cred.txt

* Revert "Added functionalities for entering AWS CLI based on user groups"

This reverts commit 8ca7cb2ef2e137986691ad74d4ab292ece41b104.

* Added test-user credentials

* Made suggested changes to reduce merge conflicts for a stable branch

* Deleted default.config

* Made the following changes for merging

* Call -run-admin

* Fixed conflicts

---------


* removed firehosesdk files and rmvd commented code in producer.rs

* saveiam

* may be a delay on aws. logs create own table

* add item and desc to columns

* switching to text menu

* rebase

* new test code starts at src/main.rs:71

* update readme with project executive summary

* added log generation script, added instruction in comments below script

* Updated to reclaim dev

* Added rust and aws cli installation to script

* added the setup terraform

* updated gitignore

* Revised install script

* edited terraform script and loggen script, added log directory

* loggen script now creates directory for logs

* merged loggen and install scripts into install.sh, added db packages to default.toml, made the .tf script add the user to users and create a cloudwatch metric data stream for their log tables

* refactored menu and linting

* Removed a stray bracket

* Moved AWS CLI function to iam.rs and include it in menu
Menu changes reflect AWS CLI

Considering deleting util function, moving away from that text menu
style

Changed install script with setting Rust enviornment

* testing running client in the background

* default.toml:
	Changed to reflect the pre-dev
config.rs:
	The user is now able to configure the toml file based on the
menu
dynamosdk.rs:
	Chanegd aspects of the code to reflect collector running in the
background
main.rs:
	Handles multiple logs and runs it in the background
producer.rs:
	use sync watch to receive the shutdown signal each time a log
starts its own channel/thread
util.rs:
	moved menu: considering on deleting util.rs

* update

* added comments

* Linting and removing unsued features
See Backlog for details on iam and util file deletion

* Minor changes such as changing the configuration file now looks for default.config rather than config.toml. This is for the sake of new users as config.toml is looked for.

* added cred and config code

* enabled debug print statements

* added debg print statements

* added dependency code at top of file

* fixed provider error

* fixed provider error

* added dependencies to the readme

---------

+++ install.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
 # Define the log directory path
-LOG_DIR="/var/log/logminer/logs"
-# TODO - add the setup to add user permissions to interact with the log directory
-# setfacl -m u:username:perms /path/to/log/file
+LOG_DIR="./logs"
+LOG_PREFIX="test"
+MAX_LOGS=3
 
-# Create the directory if it doesn't exist
+# Create the log directory if it doesn't exist
 if [ ! -d "$LOG_DIR" ]; then
     mkdir -p "$LOG_DIR"
     chmod 700 "$LOG_DIR"  # Adjust permissions as needed
@@ -14,23 +14,84 @@ else
     echo "Log directory already exists at: $LOG_DIR"
 fi
 
-# Set the desired Terraform version
-TERRAFORM_VERSION="1.2.9"
+# TODO - add the setup to add user permissions to interact with the log directory
+# Replace 'username' with the actual username that needs permissions
+# setfacl -m u:username:rwx "$LOG_DIR"
 
-# Download Terraform. Adjust the version number as necessary.
-wget https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/terraform_${TERRAFORM_VERSION}_linux_amd64.zip
 
-# Unzip the downloaded file
-unzip terraform_${TERRAFORM_VERSION}_linux_amd64.zip
+# Function to generate a random security log entry
+generate_log_entry() {
+  TIMESTAMP=$(date +"%Y-%m-%d %H:%M:%S")
+  EVENT_ID=$((RANDOM % 1000 + 1000))
+  SOURCE="Security"
+  MESSAGE="Random security log message with ID $EVENT_ID"
+  echo "$TIMESTAMP - $SOURCE - Event ID: $EVENT_ID - $MESSAGE"
+}
 
-# Move the executable to a directory included in the system's PATH
-sudo mv terraform /usr/local/bin/
+# Create and write to the log files
+for ((i=1; i<=MAX_LOGS; i++)); do
+  LOG_FILE="${LOG_DIR}/${LOG_PREFIX}${i}.log"
+  LOG_ENTRY=$(generate_log_entry)
+  echo "$LOG_ENTRY" > "$LOG_FILE"
+done
 
-# Remove the downloaded ZIP file
-rm terraform_${TERRAFORM_VERSION}_linux_amd64.zip
+echo "Log files created successfully."
 
-# Check the installation
-terraform --version
+# Check if Terraform is already installed
 
-# Print success message
-echo "Terraform installed successfully."
+if [ ! -f "/usr/local/bin/terraform" ]; then
+
+    # Set the desired Terraform version
+    TERRAFORM_VERSION="1.2.9"
+
+    # Download Terraform. Adjust the version number as necessary.
+    wget https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/terraform_${TERRAFORM_VERSION}_linux_amd64.zip
+
+    # Unzip the downloaded file
+    unzip terraform_${TERRAFORM_VERSION}_linux_amd64.zip
+
+    # Move the executable to a directory included in the system's PATH
+    sudo mv terraform /usr/local/bin/
+
+    # Remove the downloaded ZIP file
+    rm terraform_${TERRAFORM_VERSION}_linux_amd64.zip
+
+    # Check the installation
+    terraform --version
+
+    # Print success message
+    echo "Terraform installed successfully."
+
+else
+    echo "Terraform is already installed."
+    terraform --version
+fi
+
+# Check if AWS CLI is already installed
+if command -v aws >/dev/null 2>&1; then
+    echo "AWS CLI is already installed."
+    aws --version
+else
+    echo "AWS CLI is not installed. Proceeding with installation..."
+
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
+    unzip awscliv2.zip
+    sudo ./aws/install
+    rm -rf awscliv2.zip aws/
+
+fi
+
+# Check if Rust is already installed
+if command -v rustc >/dev/null 2>&1; then
+    echo "Rust is already installed."
+    rustc --version
+    cargo --version
+else
+    echo "Rust is not installed. Proceeding with installation..."
+    # Install required dependencies
+    sudo apt install -y curl build-essential
+    # Download and run the Rust installation script
+    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
+
+
+fi
